feat: get rid of emulated action
