+++ front/lib/actions/conversation/list_files.ts
@@ -1,10 +1,4 @@
-import {
-  DEFAULT_CONVERSATION_EXTRACT_ACTION_NAME,
-  DEFAULT_CONVERSATION_INCLUDE_FILE_ACTION_NAME,
-  DEFAULT_CONVERSATION_LIST_FILES_ACTION_NAME,
-  DEFAULT_CONVERSATION_QUERY_TABLES_ACTION_NAME,
-  DEFAULT_CONVERSATION_SEARCH_ACTION_NAME,
-} from "@app/lib/actions/constants";
+import { DEFAULT_CONVERSATION_LIST_FILES_ACTION_NAME } from "@app/lib/actions/constants";
 import type { ExtractActionBlob } from "@app/lib/actions/types";
 import { BaseAction } from "@app/lib/actions/types";
 import type { ConversationAttachmentType } from "@app/lib/api/assistant/conversation/attachments";
@@ -47,15 +41,7 @@ export class ConversationListFilesActionType extends BaseAction {
   }
 
   async renderForMultiActionsModel(): Promise<FunctionMessageTypeModel> {
-    let content =
-      `When a user attaches a file to the conversation, an <attachment> tag marks its position in the conversation history. ` +
-      `This tag indicates when the file was attached but does not contain its content. ` +
-      `Files attached to the conversation are listed below with their content type and status (includable, queryable, searchable):\n\n` +
-      `// includable: full content can be retrieved using \`${DEFAULT_CONVERSATION_INCLUDE_FILE_ACTION_NAME}\`\n` +
-      `// queryable: represents tabular data that can be queried alongside other queryable files' tabular data using \`${DEFAULT_CONVERSATION_QUERY_TABLES_ACTION_NAME}\`\n` +
-      `// searchable: content can be searched alongside other searchable files' content using \`${DEFAULT_CONVERSATION_SEARCH_ACTION_NAME}\`\n` +
-      `// extractable: files can also be processed to extract structured data using \`${DEFAULT_CONVERSATION_EXTRACT_ACTION_NAME}\`\n` +
-      `Other tools that accept files (referenced by their id) as arguments can be available. Rely on their description and the files mime types to decide which tool to use on which file.`;
+    let content = `The following files are currently attached to the conversation:\n`;
     for (const [i, attachment] of this.files.entries()) {
       if (i > 0) {
         content += "\n";
